@@ -5,18 +5,18 @@ requires = ["hatchling", "hatch-vcs"]
 [project]
 name = "anndata"
 description = "Annotated data."
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 license = "BSD-3-Clause"
 authors = [
-    {name = "Philipp Angerer"},
-    {name = "Alex Wolf"},
-    {name = "Isaac Virshup"},
-    {name = "Sergei Rybakov"},
+    { name = "Philipp Angerer" },
+    { name = "Alex Wolf" },
+    { name = "Isaac Virshup" },
+    { name = "Sergei Rybakov" },
 ]
 maintainers = [
-    {name = "Isaac Virshup", email = "ivirshup@gmail.com"},
-    {name = "Philipp Angerer", email = "philipp.angerer@helmholtz-munich.de"},
-    {name = "Alex Wolf", email = "f.alex.wolf@gmx.de"},
+    { name = "Isaac Virshup", email = "ivirshup@gmail.com" },
+    { name = "Philipp Angerer", email = "philipp.angerer@helmholtz-munich.de" },
+    { name = "Alex Wolf", email = "f.alex.wolf@gmx.de" },
 ]
 readme = "README.md"
 classifiers = [
